moved receiveUserFlags to flag_utils
